common/pdf: renamed to common/tom.py. only support PoissonianTOM now
