+++ tests/common/tom_test.py
@@ -1,21 +1,21 @@
 import unittest
 
-from nhe.common.pdf import Poisson
+from nhe.common.tom import PoissonTOM
 
 
-class PoissonPDFTestCase(unittest.TestCase):
+class PoissonTOMTestCase(unittest.TestCase):
     def test_non_positive_time_span(self):
-        self.assertRaises(RuntimeError, Poisson, -1)
-        self.assertRaises(RuntimeError, Poisson, 0)
+        self.assertRaises(RuntimeError, PoissonTOM, -1)
+        self.assertRaises(RuntimeError, PoissonTOM, 0)
 
     def test_get_probability(self):
-        pdf = Poisson(time_span=50)
+        pdf = PoissonTOM(time_span=50)
         self.assertEqual(pdf.get_probability(occurrence_rate=10), 1)
         aae = self.assertAlmostEqual
         aae(pdf.get_probability(occurrence_rate=0.1), 0.9932621)
         aae(pdf.get_probability(occurrence_rate=0.01), 0.39346934)
 
-        pdf = Poisson(time_span=5)
+        pdf = PoissonTOM(time_span=5)
         self.assertEqual(pdf.get_probability(occurrence_rate=8), 1)
         aae = self.assertAlmostEqual
         aae(pdf.get_probability(occurrence_rate=0.1), 0.3934693)
